Migrate tests.
